Add `SharedVar` class.
This is a simple wrapper around `Value` and `Condition` from
`multiprocessing` that increases readability and is more convenient to
use.

Our typical usage pattern is changing the `Value` and notifying
listeners on waiting `Condition`.  Instead of passing two separate
values around, we now use a single instance of  `SharedInt` or
`SharedFloat`.

# Conflicts:
#	src/neptune_scale/core/components/sync_process.py