Block in `Run.close/terminate` until completed
This applies to cases when we close a Run from
a different thread. It's better to wait until
the worker process terminates, as it makes it easier to print out a
final error message in case of an error, instead of having it mixed
together with logging from different threads.